requirements.txt
fastapi
uvicorn[standard]
pydantic
beautifulsoup4
requests
pydub
openai
python-dotenv
pandas
black
isort
pytest
